+++ README.md
@@ -1,4 +1,4 @@
 # movies-explorer-api
 Backend for movie app
 Домен, по которому можно обратиться к серверу:
-api.movies-explorerolga.nomoreparties.co
+https://api.movies-explorerolga.nomoreparties.co
