+++ app.js
@@ -5,8 +5,8 @@ const mongoose = require('mongoose');
 const { celebrate, Joi } = require('celebrate');
 const { errors } = require('celebrate');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-const helmet = require('helmet');
-const { rateLimit } = require('express-rate-limit');
+const { helmet } = require('helmet');
+const { limiter } = require('./middlewares/limiter');
 
 // eslint-disable-next-line no-undef
 const { DATABASE_URL = 'mongodb://127.0.0.1:27017/bitfilmsdb'} = process.env;
@@ -42,14 +42,6 @@ app.use(
     optionsSuccessStatus: 204,
   })
 );
-
-const limiter = rateLimit({
-	windowMs: 15 * 60 * 1000, // 15 minutes
-	max: 100, // Limit each IP to 100 requests per `window` (here, per 15 minutes)
-	standardHeaders: 'draft-7', // draft-6: RateLimit-* headers; draft-7: combined RateLimit header
-	legacyHeaders: false, // X-RateLimit-* headers
-	// store: ... , // Use an external store for more precise rate limiting
-})
 
 // Apply the rate limiting middleware to all requests
 app.use(limiter);
